Add separate ping component, disable hardware component at startup
